update
Given an input of 5 strings that each represent a playing card. Output a valid poker hand.
・A card is represented as a string that combines a char that represents a Suit and a string that represents a Rank.
　○The chars that represent a Suit are, S(=Spades♠), H(=Hearts♡), D(=Diamonds♢), C(=Clovers♣).
　○Tha strings that represent a Rank are 2, 3, …, 9, 10, J, Q, K, A.
　○For example, D3 means 3 of Diamonds. CA means A of Clovers.
・The poker hands are shown in Table 1. If the 5 cards produce no hands, the output is “--”.
・Cards don’t include a Joker and there are 52 cards in the deck.

For example, if the input string is D4C4C8D8S4, then there is a Full House, so output is FH.
if the input string is S8D3HQS3CQ, it has Two pairs. the output is 2P.
(1) Write the code that satisfies the requirements above in the language you choose and write the command how to execute the program.